requirements.txt
# # ===============================
# # CORE ML (MATCH COLAB STABLE)
# # ===============================
# tensorflow==2.15.0
# keras==2.15.0
# numpy==1.26.4
# pandas==2.2.2
# scikit-learn==1.4.2
# xgboost==2.0.3
# scipy==1.13.1
# joblib==1.4.2

# # ===============================
# # FASTAPI BACKEND
# # ===============================
# fastapi==0.111.0
# uvicorn==0.30.1
# python-multipart==0.0.9
# pydantic==2.7.4

# # ===============================
# # DATABASE (MYSQL / SQLALCHEMY)
# # ===============================
# sqlalchemy==2.0.30
# pymysql==1.1.1

# # ===============================
# # UTILITIES
# # ===============================
# python-dotenv==1.0.1
# requests==2.32.3

# # ===============================
# # OPTIONAL (RECOMMENDED)
# # ===============================
# openpyxl==3.1.5


fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
pymysql
python-jose
passlib[bcrypt]
pydantic
pydantic-settings
argon2-cffi
python-multipart
apscheduler
tensorflow
keras 
numpy
pandas
scikit-learn
xgboost
scipy
joblib
requests
openpyxl
pydantic[email]
email-validator
cryptography